Part 1 of refactor complete, need to make more modular for editing requests
